requirements.txt
certifi==2026.1.4
fastapi==0.129.2
google-genai==1.73.1
httpx==0.28.1
livekit==1.1.5
livekit-agents==1.5.5
livekit-api==1.1.0
livekit-plugins-google==1.5.5
livekit-plugins-silero==1.5.5
pytz==2025.2
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sentry-sdk==2.53.0
supabase==2.28.0
uvicorn==0.41.0
